backend/pyproject.toml
[project]
name = "backend"
version = "1.0.0"
requires-python = ">=3.10"
dependencies = [
    "alembic==1.18.4",
    "annotated-doc==0.0.4",
    "annotated-types==0.7.0",
    "anyio==4.13.0",
    "argon2-cffi>=25.1.0",
    "asyncpg==0.31.0",
    "certifi==2026.2.25",
    "click==8.3.1",
    "email-validator>=2.3.0",
    "fastapi==0.135.2",
    "greenlet==3.3.2",
    "h11==0.16.0",
    "httpcore==1.0.9",
    "httptools==0.7.1",
    "httpx==0.28.1",
    "idna==3.11",
    "mako==1.3.10",
    "markupsafe==3.0.3",
    "passlib[bcrypt]>=1.7.4",
    "pydantic==2.12.5",
    "pydantic-core==2.41.5",
    "pydantic-settings==2.13.1",
    "python-dateutil>=2.9.0.post0",
    "python-dotenv==1.2.2",
    "python-jose[cryptography]>=3.5.0",
    "python-multipart>=0.0.22",
    "pyyaml==6.0.3",
    "sqlalchemy==2.0.48",
    "starlette==1.0.0",
    "typing-extensions==4.15.0",
    "typing-inspection==0.4.2",
    "uvicorn==0.42.0",
    "uvloop==0.22.1",
    "watchfiles==1.1.1",
    "websockets==16.0",
    "xmltodict>=1.0.4",
    "pyotp>=2.9.0",
    "slowapi>=0.1.9",
]
